+++ gridstatus/epa/__init__.py
@@ -1,0 +1,3 @@
+from gridstatus.epa.epa import EPA
+
+__all__ = ["EPA"]